backend/requirements.txt
# Django and REST Framework
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.6.0

# Database
psycopg2-binary==2.9.10  # PostgreSQL adapter
python-dotenv==1.0.1  # Environment variables

# PDF Generation
reportlab==4.2.5

# Audio Processing and Speech Recognition
SpeechRecognition==3.10.4
pydub==0.25.1
PyAudio==0.2.14  # Optional, for better audio support

# AI/NLP Libraries
google-generativeai==0.8.3  # For Gemini AI interviewer (recommended)
openai==1.59.6  # For GPT-based AI interviewer (alternative)
anthropic==0.43.1  # Alternative AI provider

# Utilities
python-dotenv==1.0.1
Pillow==11.1.0
pytz==2025.1

# Development
ipython==8.31.0
